+++ packages/server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xpert-ai/server-core
 
+## 3.9.6
+
+### Patch Changes
+
+- Updated dependencies [2558760]
+    - @xpert-ai/plugin-sdk@3.9.8
+
 ## 3.9.5
 
 ### Patch Changes